fix(time): separate abstraction of `TimeConverter`

+++ crates/core/time.rs
@@ -6,6 +6,6 @@ impl DateTime {
     pub fn get_local_current() -> (String, u64) {
         let date_time = LocalTimeTool::get_date_time();
         let unix_time = LocalTimeTool::convert_to_unix_time(&date_time).unwrap_or(0);
-        (date_time, unix_time)
+        (date_time, unix_time as u64)
     }
 }
